// gcpe-MediaHub/gcpe-mediahub.client/src/api/generated-client/model/request-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * GCPE Media Hub 2.0 API
 * GCPE Media Hub 2.0
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface RequestDto
 */
export interface RequestDto {
    /**
     * 
     * @type {string}
     * @memberof RequestDto
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof RequestDto
     */
    'requestTitle'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof RequestDto
     */
    'receivedOn'?: string;
    /**
     * 
     * @type {string}
     * @memberof RequestDto
     */
    'deadline'?: string;
    /**
     * 
     * @type {string}
     * @memberof RequestDto
     */
    'leadMinistryAbbr'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof RequestDto
     */
    'requestorContactFirstName'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof RequestDto
     */
    'requestorContactLastName'?: string | null;
    /**
     * 
     * @type {number}
     * @memberof RequestDto
     */
    'requestNo'?: number;
    /**
     * 
     * @type {Array<string>}
     * @memberof RequestDto
     */
    'additionalMinistriesAbbr'?: Array<string> | null;
    /**
     * 
     * @type {string}
     * @memberof RequestDto
     */
    'assignedToFullName'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof RequestDto
     */
    'requestStatus'?: string | null;
}